feat: Plugins to NavBar
