+++ src/scss/layout.scss
@@ -13,7 +13,7 @@
     align-items: center;
   }
 
-  &.right {
+  &.end {
     justify-content: flex-end;
   }
